Automatically set up the database on first run if it's missing
