--- src/models/__init__.py
@@ -1,2 +1,0 @@
-from .investors import Investor
-from .investments import Investment
